refactor resume books
