+++ apps/admin-dashboard/app/routes/_dashboard.resume-books.create.tsx
@@ -7,6 +7,7 @@ import {
 import { Form as RemixForm, useActionData, useFetcher } from '@remix-run/react';
 import { useEffect } from 'react';
 
+import { createResumeBook } from '@oyster/core/resume-books';
 import {
   Button,
   ComboboxPopover,
@@ -24,7 +25,6 @@ import {
   validateForm,
 } from '@oyster/ui';
 
-import { createResumeBook } from '@/member-profile.server';
 import { CreateResumeBookInput } from '@/member-profile.ui';
 import { type SearchCompaniesResult } from '@/routes/api.companies.search';
 import { Route } from '@/shared/constants';
