+++ packages/core/package.json
@@ -20,8 +20,8 @@
     "./object-storage": "./src/modules/object-storage/index.ts",
     "./resources": "./src/modules/resource/index.ts",
     "./resources.server": "./src/modules/resource/index.server.ts",
-    "./resume-books": "./src/modules/resume-book/index.ts",
-    "./resume-books.server": "./src/modules/resume-book/index.server.ts",
+    "./resume-books": "./src/modules/resume-book/resume-book.core.ts",
+    "./resume-books.types": "./src/modules/resume-book/resume-book.types.ts",
     "./slack.server": "./src/modules/slack/index.server.ts"
   },
   "scripts": {
