--- packages/core/src/modules/resume-book/index.server.ts
@@ -1,5 +1,0 @@
-export { getResumeBook } from './queries/get-resume-book';
-export { getResumeBookSubmission } from './queries/get-resume-book-submission';
-export { listResumeBooks } from './queries/list-resume-books';
-export { listResumeBookSponsors } from './queries/list-resume-book-sponsors';
-export { submitResume } from './use-cases/submit-resume';
