--- packages/core/src/modules/resume-book/index.ts
@@ -1,1 +1,0 @@
-export { SubmitResumeInput } from './resume-book.types';
